ledger-service/requirements.txt
fastapi
uvicorn
psycopg2-binary
pydantic
requests
python-dotenv